requirements.txt
fastapi
sqlalchemy 
aiosqlite
python-jose[cryptography]
passlib
httpx
redis
pydantic-settings
pytest 
pytest-asyncio 
pytest-mock
